Deduplicate MDL payload decrypt/verify

+++ heidi-proximity/src/commonMain/kotlin/ch/ubique/heidi/proximity/protocol/mdl/MdlEngagement.kt
@@ -68,7 +68,7 @@ data class MdlEngagement(val coseKey: ByteArray,
 			// we could have capabilities
 			deviceEngagement[Value.Number(JsonNumber.Integer(6))]?.asOrderedObject()?.let { capabilitiesObject ->
 				val caps = mutableMapOf<Int, MdlCapability>()
-				capabilitiesObject[Value.Number(JsonNumber.Integer(MdlCapabilities.DC_API_CAPABILITY_KEY))]?.let { dcApiCapability ->
+				capabilitiesObject[Value.Number(JsonNumber.Integer(MdlCapabilities.DC_API_CAPABILITY_KEY.toLong()))]?.let { dcApiCapability ->
 					dcApiCapability.asArray()?.mapNotNull { it.asString() }?.let{
 						caps.put(MdlCapabilities.DC_API_CAPABILITY_KEY, DcApiCapability(it))
 					}
